+++ CHANGELOG.md
@@ -30,6 +30,10 @@ Contains all the PRs that improved the code without changing the behaviours.
 
 ## [v2.0.0]
 
+### Added
+
+- [#416](https://github.com/archway-network/archway/pull/416) - Enable ICAHost
+
 ### Fixed
 
 - [#414](https://github.com/archway-network/archway/pull/414) - Preventing user from setting contract flat fee if rewards address is not set